chore: remove OM API Explorer, keep route introspection endpoint (OMD-1284)
Wave 1 Cluster #20 Part C: API Explorer has been migrated to OMAI
(PR #99, OMD-1283) with dual-target support. This removes:

- server/src/api/apiExplorer.js (506 LOC — test CRUD, runner, route introspection)
- front-end/src/features/devel-tools/api-explorer/ApiExplorerPage.tsx (969 LOC)
- Menu entries from MenuItems.ts, SiteMapPage, ServerDevOpsPage,
  AdminControlPanel, SystemServerPage

Retained: server/src/api/systemRoutes.js — minimal read-only route
introspection endpoint (GET /api/system/routes) that OMAI's dual-target
API Explorer proxies to when target=om.

Route /devel-tools/api-explorer now redirects to /admin/control-panel.
Added deprecation registry entry at stage 4.
