+++ front-end/src/features/admin/control-panel/system-server/ServerDevOpsPage.tsx
@@ -33,7 +33,7 @@ const TOOLS = [
   { title: 'System Information', description: 'Node.js version, uptime, memory, CPU, and application settings', href: '/admin/settings?tab=general', icon: <SysInfoIcon /> },
   { title: 'Backup & Restore', description: 'Backup configuration, storage info, job history, and restore points', href: '/admin/settings?tab=backup', icon: <BackupIcon /> },
   { title: 'Services Monitor', description: 'System health, running services status, and process management', href: '/admin/settings?tab=services', icon: <ServicesIcon /> },
-  { title: 'API Explorer', description: 'Browse and test backend API endpoints interactively', href: '/devel-tools/api-explorer', icon: <ApiIcon /> },
+  // API Explorer — migrated to OMAI (OMD-1283)
   { title: 'Build Info', description: 'Current build version, commit hash, and deployment details', href: '/devel-tools/build-info', icon: <BuildIcon /> },
   { title: 'OMTrace Console', description: 'Real-time system tracing and debug output', href: '/devel-tools/omtrace', icon: <TraceIcon /> },
   { title: 'Refactor Console', description: 'Code refactoring tools and analysis', href: '/devel-tools/refactor-console', icon: <RefactorIcon /> },
